+++ src/main/kotlin/racingcar/InputView.kt
@@ -1,0 +1,11 @@
+package racingcar
+
+class InputView() {
+    fun drawNumberOfCarInput() {
+        println("자동차 대수는 몇 대인가요?")
+    }
+
+    fun drawNumberOfTryInput() {
+        println("시도할 횟수는 몇 회인가요?")
+    }
+}
